requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
django-ratelimit
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
